backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
scikit-learn
pandas
numpy
joblib
